MDEV-37379 Assertion `index->is_readable()' failed on REPLACE DELAYED
The assertion was failing because innodb was initialized with the trash
memory caused by incomplete LEX initialization in Delayed_indert,
particularly LEX::create_info.

We could catch that error with access to poisoned memory as well.

Tables are not used to be created in delayed insert background, so this
incompleme initialization was enough.

Solution:
1. Make a better initialization by calling lex_start in the master
thread.
2. Delayed inserts still won't work for GTTs, so forbid them.