MDEV-33640: Server crashes at my_hash_free
Analysis:
While parsing the json schema, we dont initialize the hash that stores
properties, but we try to free hash while cleanup. This happend because
the variable that checks whether hash is initialized has garbage value.
So the check becomes true and we end up deleting the uninitialized hash.

Fix:
Initialize the is_hash_inited to false.