MDEV-38295 Recursive CTE usage leaks memory when not used in a top level select
During st_select_lex::cleanup() we assume that leaf tables and any
associated recursive table references can only be present when a join
is present.  After MDEV-37220 this is no longer true.
We shift our cleanup routine out of the test for a join.
We also add Multidelete_prelocking_strategy::handle_end() to populate
these leaf_tables in the outermost select when it is a delete statement.